# collection.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "postman_collection",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "paths": {
...
    "/api/hotels/bookings": {
      "get": {
        "summary": "List Hotel Bookings",
        "description": "### Get Hotel Bookings\n\n**Genereated using PostbotAI\n\nThis endpoint retrieves the bookings for a specific hotel.\n\n#### Request\n\n- Method: GET\n    \n- URL: `http://localhost:3000/api/hotels/bookings`\n    \n- Query Parameters:\n    \n    - hotelId (required): The ID of the hotel for which bookings are to be retrieved.\n        \n\n#### Response\n\nThe response for this request is a JSON object representing the bookings for the specified hotel. The JSON schema for the response can be described as follows:\n\n``` json\n{\n  \"type\": \"object\",\n  \"properties\": {\n    \"bookingId\": {\n      \"type\": \"string\"\n    },\n    \"guestName\": {\n      \"type\": \"string\"\n    },\n    \"checkInDate\": {\n      \"type\": \"string\",\n      \"format\": \"date\"\n    },\n    \"checkOutDate\": {\n      \"type\": \"string\",\n      \"format\": \"date\"\n    },\n    \"roomType\": {\n      \"type\": \"string\"\n    },\n    \"totalAmount\": {\n      \"type\": \"number\"\n    }\n  }\n}\n\n ```",
        "operationId": "listHotelBookings",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create Hotel Bookings",
        "description": "Create Hotel Bookings",
        "operationId": "createHotelBookings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "checkInDate": {
                    "type": "string",
                    "example": "2025-03-10"
                  },
                  "checkOutDate": {
                    "type": "string",
                    "example": "2025-03-12"
                  },
                  "hotelId": {
                    "type": "number",
                    "example": 5
                  },
                  "roomTypeId": {
                    "type": "number",
                    "example": 5
                  }
                }
              },
              "examples": {
                "Create Hotel Bookings": {
                  "value": {
                    "checkInDate": "2025-03-10",
                    "checkOutDate": "2025-03-12",
                    "hotelId": 5,
                    "roomTypeId": 5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Cancel Hotel Booking",
        "description": "This endpoint is used to cancel a hotel booking by sending an HTTP DELETE request to the specified URL. The request should include the bookingId as a query parameter to identify the booking to be canceled.\n\n**Genereated using PostbotAI\n\n### Request Body\n\nThis request does not require a request body.\n\n### Response\n\nUpon a successful cancellation, the response will include a confirmation message indicating that the booking has been successfully canceled. If the bookingId provided is not found or there is an error in the cancellation process, an appropriate error message will be returned.",
        "operationId": "cancelHotelBooking",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookingId": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "examples": {
                "Cancel Hotel Booking": {
                  "value": {
                    "bookingId": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/bookings/view": {
      "get": {
        "summary": "List Bookings",
        "description": "### API Request Description\n\n**Genereated using PostbotAI\n\nThis endpoint makes an HTTP GET request to retrieve the booking details for a specific itinerary ID.\n\n#### Request\n\n- Method: GET\n    \n- URL: `http://localhost:3000/api/bookings/view`\n    \n- Query Parameter:\n    \n    - `itineraryId` (number) - The ID of the itinerary for which the booking details are to be retrieved.\n        \n\n#### Response\n\nThe response is in JSON format and includes the booking details for the specified itinerary ID. The response schema is as follows:\n\n``` json\n{\n  \"hotelBooking\": {},\n  \"flightBooking\": {\n    \"bookingReference\": \"\",\n    \"ticketNumber\": \"\",\n    \"firstName\": \"\",\n    \"lastName\": \"\",\n    \"email\": \"\",\n    \"passportNumber\": \"\",\n    \"status\": \"\",\n    \"agencyId\": \"\",\n    \"createdAt\": \"\",\n    \"flights\": [\n      {\n        \"id\": \"\",\n        \"flightNumber\": \"\",\n        \"departureTime\": \"\",\n        \"arrivalTime\": \"\",\n        \"originId\": \"\",\n        \"destinationId\": \"\",\n        \"duration\": 0,\n        \"price\": 0,\n        \"currency\": \"\",\n        \"availableSeats\": 0,\n        \"status\": \"\",\n        \"airline\": {\n          \"code\": \"\",\n          \"name\": \"\"\n        },\n        \"origin\": {\n          \"code\": \"\",\n          \"name\": \"\",\n          \"city\": \"\",\n          \"country\": \"\"\n        },\n        \"destination\": {\n          \"code\": \"\",\n          \"name\": \"\",\n          \"city\": \"\",\n          \"country\": \"\"\n        }\n      }\n    ]\n  }\n}\n\n ```",
        "operationId": "listBookings",
        "parameters": [
          {
            "name": "itineraryId",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}